requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.4.0
anthropic==0.34.2
httpx<0.28
stripe>=11,<12
celery==5.4.0
redis==5.0.8
google-auth==2.35.0
google-auth-oauthlib==1.2.1
requests==2.32.3
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.7.0
